# swagger/swagger.yaml
basePath: /
definitions:
  feeds.Feed:
    properties:
      channel:
        type: string
      epoch:
        type: integer
      extra_parameters:
        type: string
      formatted_address:
        type: string
      full_text:
        type: string
      id:
        type: integer
      is_resolved:
        type: boolean
      reason:
        type: string
      timestamp:
        type: string
    type: object
  feeds.FeedLocation:
    properties:
      address:
        type: string
      entry_id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
    type: object
  feeds.Result:
    properties:
      channel:
        type: string
      entry_id:
        type: integer
      epoch:
        type: integer
      extra_parameters:
        type: string
      id:
        type: integer
      is_location_verified:
        type: boolean
      is_need_verified:
        type: boolean
      loc:
        items:
          type: number
        type: array
      reason:
        type: string
      timestamp:
        type: string
    type: object
  feeds.UpdateFeedLocationsRequest:
    properties:
      feed_locations:
        items:
          $ref: '#/definitions/feeds.FeedLocation'
        type: array
    type: object
  handler.RawFeed:
    properties:
      channel:
        type: string
      epoch:
        type: integer
      extra_parameters:
        type: string
      id:
        type: string
      raw_text:
        type: string
    type: object
  handler.request:
    properties:
      feeds:
        items:
          $ref: '#/definitions/handler.RawFeed'
        type: array
    type: object
  needs.CreateNeedRequest:
    properties:
      address:
        type: string
      description:
        type: string
    required:
    - address
    - description
    type: object
  needs.LiteNeed:
    properties:
      id:
        type: integer
    type: object
  needs.Need:
    properties:
      description:
        type: string
      extra_parameters:
        type: string
      formatted_address:
        type: string
      id:
        type: integer
      is_resolved:
        type: boolean
      loc:
        items:
          type: number
        type: array
      timestamp:
        type: string
    type: object
  needs.Response:
    properties:
      count:
        type: integer
      results:
        items:
          $ref: '#/definitions/needs.Need'
        type: array
    type: object
host: apigo.afetharita.com
info:
  contact: {}
  description: This is a sample swagger for Afet Harita
  license:
    name: Apache License, Version 2.0 (the "License")
    url: https://github.com/acikkaynak/deprem-yardim-backend-go/blob/main/LICENSE
  title: Afet Harita API
  version: "1.0"
paths:
  /events:
    post:
      consumes:
      - application/json
      parameters:
      - description: RequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create Event areas with request body
      tags:
      - Event
  /feeds/{id}:
    get:
      parameters:
      - description: Feed Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/feeds.Feed'
      summary: Get Feeds with given id
      tags:
      - Feed
  /feeds/areas:
    get:
      parameters:
      - description: Sw Lat
        in: query
        name: sw_lat
        required: true
        type: number
      - description: Sw Lng
        in: query
        name: sw_lng
        required: true
        type: number
      - description: Ne Lat
        in: query
        name: ne_lat
        required: true
        type: number
      - description: Ne Lng
        in: query
        name: ne_lng
        required: true
        type: number
      - description: Timestamp
        in: query
        name: time_stamp
        type: integer
      - description: Reason
        in: query
        name: reason
        type: string
      - description: Channel
        in: query
        name: channel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feeds.Result'
            type: array
      summary: Get Feed areas with query strings
      tags:
      - Feed
    patch:
      consumes:
      - application/json
      parameters:
      - description: RequestBody
        in: body
        name: UpdateFeedLocationsRequest
        required: true
        schema:
          $ref: '#/definitions/feeds.UpdateFeedLocationsRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
      security:
      - ApiKeyAuth: []
      summary: Update feed locations with correct address and location
      tags:
      - Feed
  /healthcheck:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Show the status of server.
      tags:
      - HealthCheck
  /needs:
    get:
      parameters:
      - description: Is Only Not Resolved
        in: query
        name: only_not_resolved
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/needs.Response'
      summary: Get Needs
      tags:
      - Need
    post:
      parameters:
      - description: RequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/needs.CreateNeedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/needs.LiteNeed'
      security:
      - ApiKeyAuth: []
      summary: Create Need
      tags:
      - Need
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-Api-Key
    type: apiKey
swagger: "2.0"